requirements.txt
yahoo-fantasy-api>=1.9.0
yahoo-oauth>=1.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
nba_api>=1.5.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
